updateCheck clean!
